overwrite old FormatType
